fix: An error occurred while creating the project because manifest.json does not exist

+++ packages/core/src/utils/setPackageName.ts
@@ -1,7 +1,9 @@
-import { readFileSync, writeFileSync } from 'node:fs'
+import { existsSync, readFileSync, writeFileSync } from 'node:fs'
 import { join } from 'node:path'
 
 function replaceNameContent(filePath: string, projectName: string) {
+  if (!existsSync(filePath))
+    return
   const fileContent = JSON.parse(readFileSync(filePath, 'utf8'))
   fileContent.name = projectName
   writeFileSync(filePath, JSON.stringify(fileContent, null, 2))
